+++ client/src/grid/PicksGrid.test.js
@@ -6,6 +6,9 @@ import {useQuery, useMutation} from '@apollo/react-hooks';
 import {mockQueryData} from "../testUtilities/MockQueryData";
 
 import gql from 'graphql-tag';
+import {assertAllUserPicksMatchCellText, findByClassName} from "../testUtilities/Helpers";
+import PickCell from "./PickCell";
+import {getPickByGame} from "./PickCells";
 
 
 jest.mock('@apollo/react-hooks');
@@ -82,23 +85,7 @@ describe('PicksGrid', () => {
         expect(useMutation.mock.calls[0][0]).toBe(updatingQuery);
     });
 
-    it('PickCells is given sendData callback', () => {
-        let grid = null;
-        const callback = () => {
-        };
-        useMutation.mockReturnValue([callback]);
-        act(() => {
-            grid = create(<PicksGrid/>)
-        });
 
-        const cell = grid.root.find(el => el.props.id === "pick-cells");
-        expect(cell.props.sendData).toBe(callback);
-    });
-
-    it('PickCells is given current week', () => {
-        const cell = grid.findByProps({id: "pick-cells"});
-        expect(cell.props.currentWeek).toBe(mockQueryData.weeks[0].name);
-    });
 
     describe('advance week', () => {
         let week0Change;
@@ -239,6 +226,215 @@ describe('PicksGrid', () => {
             expect(refetchSpy).not.toHaveBeenCalled();
         });
     })
+    describe('rendered picks cells', () => {
+        let grid, renderer;
 
+        beforeEach(() => {
+            jest.resetAllMocks();
+            useQuery.mockReturnValue({loading: false, error: null, data: mockQueryData});
+            useMutation.mockReturnValue([() => {}]);
+
+            renderer = create(<PicksGrid defaultWeek="0"/>);
+            grid = renderer.root;
+        });
+
+        it('PickCells is given sendData callback', () => {
+            let grid = null;
+            const callback = () => {
+            };
+            useMutation.mockReturnValue([callback]);
+            act(() => {
+                grid = create(<PicksGrid defaultWeek="0"/>)
+            });
+
+            const cell = grid.root.find(el => el.props.id === "pick-cells");
+            expect(cell.props.sendData).toBe(callback);
+        });
+
+        it('PickCells is given default week on initialization', () => {
+            const cell = grid.findByProps({id: "pick-cells"});
+            expect(cell.props.currentWeek).toBe(grid.props.defaultWeek);
+        });
+
+        it('Renders twelve pick cells when there are three users and four games in data response', () => {
+            const pickCells = findByClassName(grid, 'pick-cell');
+
+            expect(pickCells.length).toBe(mockQueryData.games.length * mockQueryData.users.length);
+
+            assertAllUserPicksMatchCellText(mockQueryData, pickCells);
+        });
+
+        it('Pick cells are of PickCell type', () => {
+            const pickCells = findByClassName(grid, 'pick-cell');
+
+            pickCells.map(cell => expect(cell.type).toBe(PickCell))
+        });
+
+        it('can choose specific game from pick list for first mock user', () => {
+            let picks = mockQueryData["userPicks"][0].picks;
+
+            expect(getPickByGame(picks, "CHI@GB")).toBe("CHI")
+        });
+
+        it('can choose specific game from pick list for second mock user', () => {
+            let picks = mockQueryData["userPicks"][1].picks;
+
+            expect(getPickByGame(picks, "ANN@COL")).toBe("C")
+        });
+
+        it('empty list of picks returns null', () => {
+            let emptyPicks = [];
+
+            expect(getPickByGame(emptyPicks, "ANN@COL")).toBe(null)
+        })
+    });
+
+    describe('rendered users cells', () => {
+
+        it('Renders three id cells when there are three users in data response', () => {
+            const nameCells = findByClassName(grid, 'name-cell');
+
+            expect(nameCells.length).toBe(mockQueryData.users.length);
+            expect(nameCells.map(cell => cell.props.children))
+                .toEqual(mockQueryData.users.map(user => user.name))
+        });
+
+        it('Renders two id cells when there are two users in data response', () => {
+            const twoMockUserData = {
+                "users": [
+                    {"name": "Someone"},
+                    {"name": "Derp"},
+                ],
+                "userTotals":[],
+                "games": []
+            };
+            useQuery.mockReturnValue({loading: false, error: null, data: twoMockUserData});
+
+            const grid = create(<PicksGrid/>).root;
+            const nameCells = findByClassName(grid, 'name-cell');
+
+            expect(nameCells.length).toBe(twoMockUserData.users.length);
+            expect(nameCells.map(cell => cell.props.children))
+                .toEqual(twoMockUserData.users.map(user => user.name))
+        });
+    });
+
+    describe('rendered totals cells', () => {
+        it('Renders three total cells when there are three users in data response', () => {
+            const totalCells = findByClassName(grid, 'total-cell');
+
+            expect(totalCells.length).toBe(mockQueryData.users.length);
+            expect(totalCells.map(cell => cell.props.children))
+                .toEqual(mockQueryData.userTotals.map(user => user.total))
+        });
+
+        it('Renders two total cells when there are two users in data response', () => {
+            const twoMockUserData = {
+                "users": [
+                    {"name": "Someone"},
+                    {"name": "Derp"},
+                ],
+                "userTotals":[
+                    {"name": "Someone", "total": 0},
+                    {"name": "Derp", "total": 4}
+                ],
+                "games": []
+            };
+            useQuery.mockReturnValue({loading: false, error: null, data: twoMockUserData});
+
+            const grid = create(<PicksGrid/>).root;
+            const totalCells = findByClassName(grid, 'total-cell');
+
+            expect(totalCells.length).toBe(twoMockUserData.users.length);
+            expect(totalCells.map(cell => cell.props.children))
+                .toEqual(twoMockUserData.userTotals.map(user => user.total))
+        });
+    });
+
+    describe('rendered games cells', () => {
+        it('Renders four game cells when there are four games in data response', () => {
+            const gameCells = findByClassName(grid, 'game-cell');
+
+            expect(gameCells.length).toBe(mockQueryData.games.length);
+            expect(gameCells.map(cell => cell.props.children))
+                .toEqual(mockQueryData.games.map(game => game.name))
+        });
+
+        it('Renders one game cell when there is one game in data response', () => {
+            const oneMockGameData = {
+                "users": [],
+                "games": [
+                    {"name": "TLH@PCL"},
+                ],
+                "userTotals":[],
+            };
+            useQuery.mockReturnValue({loading: false, error: null, data: oneMockGameData});
+
+            const grid = create(<PicksGrid/>).root;
+            const gameCells = findByClassName(grid, 'game-cell');
+
+            expect(gameCells.length).toBe(oneMockGameData.games.length);
+            expect(gameCells.map(cell => cell.props.children))
+                .toEqual(oneMockGameData.games.map(game => game.name))
+        });
+
+    });
+
+    describe('rendered spread cells', () => {
+        it('Renders four spread cells when there are four games in data response', () => {
+            const spreadCells = findByClassName(grid, 'spread-cell');
+
+            expect(spreadCells.length).toBe(mockQueryData.games.length);
+            expect(spreadCells.map(cell => cell.props.children))
+                .toEqual(mockQueryData.games.map(game => game.spread))
+        });
+
+        it('Renders one spread cell when there is one game in data response', () => {
+            const oneMockGameData = {
+                "users": [],
+                "games": [
+                    {"name": "TLH@PCL", "spread": "-20"},
+                ],
+                "userTotals":[],
+            };
+            useQuery.mockReturnValue({loading: false, error: null, data: oneMockGameData});
+
+            const grid = create(<PicksGrid/>).root;
+            const spreadCells = findByClassName(grid, 'spread-cell');
+
+            expect(spreadCells.length).toBe(oneMockGameData.games.length);
+            expect(spreadCells.map(cell => cell.props.children))
+                .toEqual(oneMockGameData.games.map(game => game.spread))
+        });
+    });
+
+    describe('rendered result cells', () => {
+        it('Renders four result cells when there are four games in data response', () => {
+            const resultCells = findByClassName(grid, 'result-cell');
+
+            expect(resultCells.length).toBe(mockQueryData.games.length);
+            expect(resultCells.map(cell => cell.props.children))
+                .toEqual(mockQueryData.games.map(game => game.result))
+        });
+
+        it('Renders one result cell when there is one game in data response', () => {
+            const oneMockGameData = {
+                "users": [],
+                "games": [
+                    {"name": "TLH@PCL", "spread": "-20", "result": "PCL"},
+                ],
+                "userTotals":[]
+            };
+            useQuery.mockReturnValue({loading: false, error: null, data: oneMockGameData});
+
+            const grid = create(<PicksGrid/>).root;
+            const resultCells = findByClassName(grid, 'result-cell');
+
+            expect(resultCells.length).toBe(oneMockGameData.games.length);
+            expect(resultCells.map(cell => cell.props.children))
+                .toEqual(oneMockGameData.games.map(game => game.result))
+        });
+
+    });
 
 });
